Add disclaimer dropdown
