+++ components/General/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
 
 export const Box: React.FC<{ className?: string }> = styled.div`
@@ -129,3 +129,89 @@ export const Logo: React.FC<LProps> = styled(({ className, ticker, clear }: LPro
     width: 30px;
     margin: 5px 0;
 `;
+
+export const Input = styled.input`
+    font-size: 42px;
+    letter-spacing: 0px;
+    color: #ffffff;
+    width: 100%;
+    &::placeholder { /* Chrome, Firefox, Opera, Safari 10.1+ */
+        color: #fff;
+        opacity: 1; /* Firefox */
+    }
+
+    &:-ms-input-placeholder { /* Internet Explorer 10-11 */
+        color: #fff;
+    }
+
+    &::-ms-input-placeholder { /* Microsoft Edge */
+        color: #fff;
+    }
+
+    &:focus {
+        border: none;
+        outline: none;
+        box-shadow: none;
+    }
+`;
+
+export const BasicInputContainer = styled.div`
+    width: 100%;
+    display: flex;
+    border-bottom: 1px solid #002886;
+`;
+
+/**
+ * Takes two children items, will place the first as the header component and the second as the body
+ * @param defaultHeight prevents jumpiness when initialising the dropdown
+ */
+export const Dropdown = styled(({ className, children, defaultHeight }) => {
+    const [open, setOpen] = useState(false);
+    const main = useRef(null);
+    const header = useRef(null);
+    const body = useRef(null);
+    useEffect(() => {
+        const h = header.current as unknown as HTMLDivElement;
+        const b = body.current as unknown as HTMLDivElement;
+        if (open) {
+            // all heights plus 10px for padding
+            (main.current as unknown as HTMLDivElement).style.height = `${h.clientHeight + b.clientHeight + 10}px`
+        } else {
+            (main.current as unknown as HTMLDivElement).style.height = `${h.clientHeight ?? defaultHeight}px`
+        }
+    }, [open])
+    return (
+        <div className={`${className} ${open ? 'open' : ''}`} onClick={(_e) => setOpen(!open)} ref={main}>
+            <div ref={header}>
+                {children[0]}
+            </div>
+            <div className="body" ref={body}>
+                {children[1]}
+            </div>
+        </div>
+    )
+})`
+    overflow: hidden;
+    transition: 0.3s ease-in-out;
+    margin-left: -10px;
+    height: ${(props) => props.defaultHeight}px;
+    margin-bottom: 2rem;
+    border-radius: 5px;
+    text-align: left;
+    font-size: 16px;
+    letter-spacing: -0.32px;
+
+    &:hover {
+        background: #03065E;
+    }
+
+    &.open > .body {
+        transition: 0.3s ease-in;
+        opacity: 0
+    }
+
+    &.open .body {
+        opacity: 1;
+    }
+`
+
