+++ components/Insurance/PoolContributors.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import Table, { TableRow } from '@components/Tables';
 import { SubNavBar } from '@components/Nav';
-import { InsuranceButtons } from '@components/Buttons/InsuranceButtons';
 import { Holding } from '@components/types';
 import LightWeightChart from '@components/Charts/LightWeightChart';
 
@@ -69,7 +68,6 @@ const PoolContributors: React.FC<PCProps> = ({ holdings, liquidity }: PCProps) =
                     </span>
                 </div>
                 <div className="ml-auto p-5 sm:w-full xl:w-1/2 flex">
-                    <InsuranceButtons />
                 </div>
             </div>
         </div>
